templates: add cluster credentials secret
Adds a credentials Secret in the cluster templates. This propagates the $LINODE_TOKEN
environment variable into a Secret that will be created for every workload cluster. While this
credentials Secret may be referenced by multiple workload clusters (or not at all), generating
a per-cluster Secret works best in the current cluster templating workflow.